# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mock EPL Server
  description: "This is the API server documentation for our mock EPL API."
  contact:
    email: barry.jonathanks@gmail.com
  version: 0.0.1
...
servers:
  - url: ""
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: "8080"
        description: The service port at which the API listens
...
paths:
...
  /fixtures/:
    post:
      operationId: create_fixture
      requestBody:
        $ref: "#/components/requestBodies/fixture_info"
      responses:
        201:
          $ref: "#/components/responses/fixture"
        400:
          $ref: "#/components/responses/bad_request"
        401:
          $ref: "#/components/responses/unauthorized"
        422:
          $ref: "#/components/responses/unprocessible_entity"
      security:
        - bearer: []
      summary: Add a new fixture (admins only)
      tags:
        - fixtures

    get:
      parameters:
        - name: status
          in: query
          schema:
            enum:
              - completed
              - pending
      operationId: list_fixtures
      responses:
        200:
          $ref: "#/components/responses/fixtures_list"
        401:
          $ref: "#/components/responses/unauthorized"
      security:
        - bearer: []
      summary: List available fixtures (requires authentication)
      tags:
        - fixtures
  /fixtures/{fixture_id}:
    parameters:
      - name: fixture_id
        in: path
        schema:
          type: string
        required: true

    delete:
      operationId: remove_fixture
      responses:
        200:
          description: Fixture removed.
        401:
          $ref: "#/components/responses/unauthorized"
      security:
        - bearer: []
      summary: Remove fixture (admins only)
      tags:
        - fixtures

    patch:
      description: Update fixture info (restricted to admins)
      operationId: update_fixture
      requestBody:
        $ref: "#/components/requestBodies/fixture_info"
      responses:
        200:
          $ref: "#/components/responses/fixture"
        400:
          $ref: "#/components/responses/bad_request"
        401:
          $ref: "#/components/responses/unauthorized"
        422:
          $ref: "#/components/responses/unprocessible_entity"
      security:
        - bearer: []
      summary: Update fixture info (admins only)
      tags:
        - fixtures
...
components:
  requestBodies:
    fixture_info:
      content:
        application/json:
          schema:
            properties:
              home_team:
                description: The ID of the home team
                type: string
              away_team:
                description: The ID of the away team
                type: string
              match_date:
                description: The date of the match
                type: string
                format: date-time
      required: true
  schemas:
...
    TeamInfo:
      properties:
        home_stadium:
          type: string
        logo_url:
          type: string
        name:
          type: string
        short_name:
          type: string
      required:
        - home_stadium
        - logo_url
        - name
        - short_name
    Team:
      description: A competitor in the EPL league.
      allOf:
        - $ref: "#/components/schemas/_Entity"
        - $ref: "#/components/schemas/TeamInfo"
    Fixture:
      description: A match arrangement between teams.
      allOf:
        - $ref: "#/components/schemas/_Entity"
        - properties:
            home_team:
              $ref: "#/components/schemas/Team"
            away_team:
              $ref: "#/components/schemas/Team"
            match_date:
              type: string
              format: date-time
    _DataResponse:
      description: An API response containing data.
      properties:
        "@type":
          type: string
          description: The resource type returned.
        message:
          type: string
          description: Information about the response.
        data: {}

      required:
        - "@type"
        - data
    _Entity:
      description: Default fields for stable entities.
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    ValidationError:
      type: object
      allOf:
        - $ref: "#/components/schemas/Error"
        - properties:
            details:
              type: array
              items:
                type: object
                properties:
                  message:
                    type: string
                  target:
                    type: string
                required:
                  - message
                  - target
  responses:
...
    fixture:
      description: Fixture information
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/_DataResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/Fixture"
                  "@type":
                    enum:
                      - "Fixture"
    fixtures_list:
      description: Fixtures list
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/_DataResponse"
              - properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Fixture"
                  "@type":
                    enum:
                      - "Fixtures"
...
    bad_request:
      description: Request malformed.
      content:
        application/json:
          schema:
            type: object
            allOf:
              - $ref: "#/components/schemas/Error"
...
    unauthorized:
      description: Authentication information provided is incorrect.
      content:
        application/json:
          schema:
            type: object
            allOf:
              - $ref: "#/components/schemas/Error"
              - properties:
                  code:
                    enum:
                      - auth/unauthorized
    unprocessible_entity:
      description: The request failed vaildation.
      content:
        application/json:
          schema:
            type: object
            allOf:
              - $ref: "#/components/schemas/ValidationError"
  securitySchemes:
    api_key:
      type: apiKey
      name: bearer
      in: header